requirements.txt
aiohttp>=3.13.2
python-telegram-bot>=22.5